BigDecimal swithc and toString added

+++ src/main/java/Homestay/enums/TIERS.java
@@ -1,17 +1,19 @@
 package main.java.Homestay.enums;
 
+import java.math.BigDecimal;
+
 public enum TIERS {
     STANDARD("Standard", 30),
     DELUXE("Deluxe", 52),
     COTTAGE("Cottage", 65);
     private final String type;
-    private final int fare;
+    private final BigDecimal fare;
     TIERS(String type, int fare){
         this.type = type;
-        this.fare = fare;
+        this.fare = BigDecimal.valueOf(fare);
     }
 
-    public int getFare() {
+    public BigDecimal getFare() {
         return fare;
     }
